pyproject.toml
[project]
name = "genai-25-sept"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.31.1",
    "ipykernel>=6.30.1",
    "openai>=1.107.3",
    "python-dotenv>=1.1.1",
]
